Fix Build Errors for .NET MAUI Preview 11
