+++ src/CommunityToolkit.Maui.Core/Views/PaddedButton.macios.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿using ObjCRuntime;
 using UIKit;
 
 namespace CommunityToolkit.Maui.Core.Views;
