+++ src/CommunityToolkit.Maui.Core/Views/Snackbar/SnackbarView.macios.cs
@@ -1,5 +1,6 @@
 using System;
 using CoreGraphics;
+using ObjCRuntime;
 using UIKit;
 
 namespace CommunityToolkit.Maui.Core.Views;
